@@ -16,24 +16,24 @@
     "@skeletonlabs/skeleton": "^2.6.0",
     "@skeletonlabs/tw-plugin": "^0.4.0",
     "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/enhanced-img": "^0.3.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/enhanced-img": "^0.4.4",
+    "@sveltejs/kit": "^2.15.2",
+    "@sveltejs/vite-plugin-svelte": "^5.0.0",
     "@tailwindcss/container-queries": "^0.1.1",
-    "@tailwindcss/forms": "0.5.7",
+    "@tailwindcss/forms": "^0.5.7",
     "autoprefixer": "^10.4.17",
     "eslint": "^9.1.1",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-svelte": "^2.30.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-svelte": "^2.45.1",
     "postcss": "^8.4.32",
-    "prettier": "^3.0.0",
-    "prettier-plugin-svelte": "^3.0.0",
-    "svelte": "^4.2.7",
+    "prettier": "^3.1.0",
+    "prettier-plugin-svelte": "^3.2.6",
+    "svelte": "^5.0.0",
     "svelte-check": "^4.0.2",
     "svelte-virtual-scroll-list": "^1.3.0",
     "tailwindcss": "^3.3.6",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.0",
+    "typescript": "^5.5.0",
+    "vite": "^6.0.7",
     "vite-plugin-tailwind-purgecss": "^0.3.3"
   },
   "type": "module",
@@ -45,5 +45,10 @@
     "minisearch": "^7.0.0",
     "openseadragon": "^5.0.0",
     "sander": "^0.6.0"
-  }
+  },
+  "description": "This website is a purely static build with a client-side search component. It is hosted by Github and was developed by [Sebastian Flick](https://github.com/flicksolutions) and [Peter Dängeli](https://github.com/pdaengeli) of the [Data Science Lab](https://github.com/dsl-unibe-ch) using Svelte and Svelte-Kit.",
+  "main": "svelte.config.js",
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
