@@ -19,10 +19,10 @@
 			src="$lib/assets/IMG_8229.jpg"
 			alt="herbarium"
 			class="max-w-full border-8 border-primary-500 anchor cursor-pointer"
-			on:click={(e) => {
+			onclick={(e) => {
 				openLightbox(e.target);
 			}}
-			on:keypress={(e) => {
+			onkeypress={(e) => {
 				if (e.key === 'Enter' || e.key === ' ') {
 					openLightbox(e.target);
 				}
@@ -42,10 +42,10 @@
 			src="$lib/assets/P1770528.jpg"
 			alt="herbarium"
 			class="max-w-full border-8 border-primary-500 anchor cursor-pointer"
-			on:click={(e) => {
+			onclick={(e) => {
 				openLightbox(e.target);
 			}}
-			on:keypress={(e) => {
+			onkeypress={(e) => {
 				if (e.key === 'Enter' || e.key === ' ') {
 					openLightbox(e.target);
 				}
@@ -74,10 +74,10 @@
 			src="$lib/assets/P1006420.jpg"
 			alt="herbarium"
 			class="max-w-full border-8 border-primary-500 anchor cursor-pointer"
-			on:click={(e) => {
+			onclick={(e) => {
 				openLightbox(e.target);
 			}}
-			on:keypress={(e) => {
+			onkeypress={(e) => {
 				if (e.key === 'Enter' || e.key === ' ') {
 					openLightbox(e.target);
 				}
